requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.1.0
boltons==23.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
config-client==1.4.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.4
Django==5.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.5
django-jazzmin==2.6.0
django-pytest==0.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
dnspython==2.4.2
drf-yasg==1.21.7
face==20.1.1
Faker==12.0.1
glom==23.5.0
idna==3.6
ifaddr==0.2.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mixer==7.2.2
openapi-codec==1.3.2
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
py-eureka-client==0.11.8
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-json-report==1.5.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
